Move PolylineArgs and MeshArgs out of internal/
