+++ core/frontend/src/core-frontend.ts
@@ -86,9 +86,11 @@ export * from "./render/FeatureSymbology";
 export * from "./render/FrameStats";
 export * from "./render/GraphicBranch";
 export * from "./render/GraphicBuilder";
+export * from "./render/MeshArgs";
 export * from "./render/MockRender";
 export * from "./render/ParticleCollectionBuilder";
 export * from "./render/Pixel";
+export * from "./render/PolylineArgs";
 export * from "./render/RealityMeshGraphicParams";
 export * from "./render/RealityMeshParams";
 export * from "./render/RenderClipVolume";
