+++ core/frontend/src/test/render/GraphicBuilder.test.ts
@@ -15,11 +15,11 @@ import { RenderSystem } from "../../render/RenderSystem";
 import { ScreenViewport } from "../../Viewport";
 import { MeshParams } from "../../common/internal/render/MeshParams";
 import { SurfaceType } from "../../common/internal/render/SurfaceParams";
-import { MeshArgs } from "../../common/internal/render/MeshPrimitives";
 import { MeshGraphic } from "../../render/webgl/Mesh";
 import { InstancedGraphicParams } from "../../common";
 import { openBlankViewport } from "../openBlankViewport";
 import { GraphicType } from "../../common/render/GraphicType";
+import { MeshArgs } from "../../render/MeshArgs";
 
 describe("GraphicBuilder", () => {
   let imodel: IModelConnection;
