+++ core/frontend/src/test/render/webgl/SurfaceTransparency.test.ts
@@ -18,9 +18,9 @@ import { Primitive } from "../../../render/webgl/Primitive";
 import { Pass, RenderPass, SinglePass } from "../../../render/webgl/RenderFlags";
 import { MeshGraphic } from "../../../render/webgl/Mesh";
 import { SurfaceGeometry } from "../../../render/webgl/SurfaceGeometry";
-import { MeshArgs } from "../../../common/internal/render/MeshPrimitives";
 import { createBlankConnection } from "../../createBlankConnection";
 import { createMeshParams } from "../../../common/internal/render/VertexTableBuilder";
+import { MeshArgs } from "../../../render/MeshArgs";
 
 function createMesh(transparency: number, mat?: RenderMaterial | RenderTexture): RenderGraphic {
   const colors = new ColorIndex();
